[web] Extract NFT fetching logic  to `useNFTBalances`
